Don't duplicate CrossJava logic
As we can now take that from upstream sbt 1.3.x

Refs #25830
